Move API calls to queries_repo.go
