+++ pkg/cmd/repo/gitignore/list/list.go
@@ -4,9 +4,9 @@ import (
 	"fmt"
 	"net/http"
 
+	"github.com/cli/cli/v2/api"
 	"github.com/cli/cli/v2/internal/gh"
 	"github.com/cli/cli/v2/internal/tableprinter"
-	"github.com/cli/cli/v2/pkg/cmd/repo/shared/queries"
 	"github.com/cli/cli/v2/pkg/cmdutil"
 	"github.com/cli/cli/v2/pkg/iostreams"
 	"github.com/spf13/cobra"
@@ -59,7 +59,7 @@ func listRun(opts *ListOptions) error {
 	}
 
 	hostname, _ := cfg.Authentication().DefaultHost()
-	gitIgnoreTemplates, err := queries.ListGitIgnoreTemplates(client, hostname)
+	gitIgnoreTemplates, err := api.ListGitIgnoreTemplates(client, hostname)
 	if err != nil {
 		return err
 	}
