+++ pkg/cmd/repo/license/list/list.go
@@ -7,7 +7,6 @@ import (
 	"github.com/cli/cli/v2/api"
 	"github.com/cli/cli/v2/internal/gh"
 	"github.com/cli/cli/v2/internal/tableprinter"
-	"github.com/cli/cli/v2/pkg/cmd/repo/shared/queries"
 	"github.com/cli/cli/v2/pkg/cmdutil"
 	"github.com/cli/cli/v2/pkg/iostreams"
 	"github.com/spf13/cobra"
@@ -60,7 +59,7 @@ func listRun(opts *ListOptions) error {
 	}
 
 	hostname, _ := cfg.Authentication().DefaultHost()
-	licenseTemplates, err := queries.ListLicenseTemplates(client, hostname)
+	licenseTemplates, err := api.ListLicenseTemplates(client, hostname)
 	if err != nil {
 		return err
 	}
